move event typedef into types directory
